Allow flexibility in getLabeledReactantsAndProducts
This method asks for a list of reactants and products in order
to output labeled ones that match the template. However, if your
input reactions list is flipped, it causes an error. This commit
fixes this bug, by trying to check other template matching if the
original order does not work.